Port more bindings from moonlander
Reorganize the bottom row to have the arrows either side of the
raise/lower buttons. Also try to put all the remaining punctuation
around the outside. However I needed to make some compromises: the
backslash is on the left, just inside the quote, and the brackets are
1 column over from on the moonlander.

I also reorganized the raise and lower layers to put F-keys on the
number row