Use layer-tap for the inner arrows to improve layout of bottom row
This is now more similar to the moonlander, with backslash in lower
right corner. Another advantage is that arrows are now snugly abutting
the space bar. The only DISadvantage is that the right and up arrows
are a bit "catchy" - to get them to do key-repeat it is necessary to
do a quick tap-then-hold. This is unavoidable if I want them to change
layers on hold, and it doesn't seem too bad.