Bring Preonic into line with Moonlander
Nove the arrow keys on the bottom row out one, and no longer share
with the raise/lower keys

Add Home/End/PgUp/PgDn as the raised or lowered arrow keys

Move the spanish accent layer activation to hold of Z and /, with the
possibility to use shift to get the upper case layer. This does
require that the layer is activated before the shift.

Add extra keys for accessing some punctuation keys (quote and
backslash) on raise and lower layers. Now we do not have to use the
outer column for any punctuation.